Add code in start.rb file
